NODE-197, feat: migrate to bp-cccp
