+++ pallets/btc-socket-queue/Cargo.toml
@@ -29,6 +29,7 @@ sp-io = { workspace = true }
 # Bifrost
 bp-btc-relay = { workspace = true }
 bp-staking = { workspace = true }
+bp-cccp = { workspace = true }
 
 # Frontier
 pallet-evm = { workspace = true }
@@ -42,23 +43,24 @@ pallet-timestamp = { workspace = true }
 [features]
 default = ["std"]
 std = [
-    "scale-info/std",
-    "parity-scale-codec/std",
-    "frame-support/std",
-    "frame-system/std",
-    "sp-runtime/std",
-    "sp-std/std",
-    "sp-core/std",
-    "sp-io/std",
-    "bp-btc-relay/std",
-    "pallet-evm/std",
+	"scale-info/std",
+	"parity-scale-codec/std",
+	"frame-support/std",
+	"frame-system/std",
+	"sp-runtime/std",
+	"sp-std/std",
+	"sp-core/std",
+	"sp-io/std",
+	"bp-btc-relay/std",
+	"bp-cccp/std",
+	"pallet-evm/std",
 ]
 try-runtime = ["frame-support/try-runtime"]
 runtime-benchmarks = [
-    "frame-benchmarking",
-    "frame-support/runtime-benchmarks",
-    "frame-system/runtime-benchmarks",
-    "sp-runtime/runtime-benchmarks",
-    "bp-btc-relay/runtime-benchmarks",
-    "pallet-balances/runtime-benchmarks",
+	"frame-benchmarking",
+	"frame-support/runtime-benchmarks",
+	"frame-system/runtime-benchmarks",
+	"sp-runtime/runtime-benchmarks",
+	"bp-btc-relay/runtime-benchmarks",
+	"pallet-balances/runtime-benchmarks",
 ]
