+++ primitives/btc-relay/src/traits.rs
@@ -76,11 +76,6 @@ pub trait SocketQueueManager<AccountId> {
 	fn set_max_fee_rate(rate: u64);
 }
 
-pub trait SocketVerifier<AccountId> {
-	/// Verify a Socket message whether it is valid.
-	fn verify_socket_message(msg: &UnboundedBytes) -> Result<(), DispatchError>;
-}
-
 pub trait BlazeManager<T: frame_system::Config> {
 	/// Check if BLAZE is activated.
 	fn is_activated() -> bool;
