+++ primitives/cccp/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "bp-cccp"
+version = "0.0.1"
+description = "Primitives for CCCP (Cross-Chain Communication Protocol)"
+authors = { workspace = true }
+homepage = { workspace = true }
+edition = { workspace = true }
+license = { workspace = true }
+repository = { workspace = true }
+
+[dependencies]
+scale-info = { workspace = true }
+parity-scale-codec = { workspace = true }
+ethabi-decode = { workspace = true }
+
+sp-std = { workspace = true }
+sp-core = { workspace = true }
+sp-runtime = { workspace = true }
+
+[features]
+default = ["std"]
+std = [
+	"scale-info/std",
+	"parity-scale-codec/std",
+	"sp-std/std",
+	"sp-core/std",
+	"sp-runtime/std",
+]
+runtime-benchmarks = []
